+++ torchmetrics/functional/audio/pit.py
@@ -101,10 +101,10 @@ def _find_best_perm_by_exhuastive_method(
     return best_metric, best_perm  # shape [batch], shape [batch, spk]
 
 
-def pit(
+def permutation_invariant_training(
     preds: torch.Tensor, target: torch.Tensor, metric_func: Callable, eval_func: str = "max", **kwargs: Dict[str, Any]
 ) -> Tuple[Tensor, Tensor]:
-    """Permutation invariant training (PIT). The PIT implements the famous Permutation Invariant Training method.
+    """Permutation invariant training. The permutation_invariant_training implements the famous Permutation Invariant Training method.
 
     [1] in speech separation field in order to calculate audio metrics in a permutation invariant way.
 
@@ -131,7 +131,7 @@ def pit(
         >>> # [batch, spk, time]
         >>> preds = torch.tensor([[[-0.0579,  0.3560, -0.9604], [-0.1719,  0.3205,  0.2951]]])
         >>> target = torch.tensor([[[ 1.0958, -0.1648,  0.5228], [-0.4100,  1.1942, -0.5103]]])
-        >>> best_metric, best_perm = pit(preds, target, scale_invariant_signal_distortion_ratio, 'max')
+        >>> best_metric, best_perm = permutation_invariant_training(preds, target, scale_invariant_signal_distortion_ratio, 'max')
         >>> best_metric
         tensor([-5.1091])
         >>> best_perm
@@ -183,7 +183,7 @@ def pit_permutate(preds: Tensor, perm: Tensor) -> Tensor:
 
     Args:
         preds (Tensor): the estimates you want to permutate, shape [batch, spk, ...]
-        perm (Tensor): the permutation returned from pit, shape [batch, spk]
+        perm (Tensor): the permutation returned from permutation_invariant_training, shape [batch, spk]
 
     Returns:
         Tensor: the permutated version of estimate